Rename calculation_X to core_calculation_X files
