+++ src/DRAMPower/DRAMPower/standards/ddr4/core_calculation_DDR4.h
