+++ src/DRAMPower/DRAMPower/standards/ddr5/core_calculation_DDR5.cpp
@@ -1,4 +1,4 @@
-#include "calculation_DDR5.h"
+#include "core_calculation_DDR5.h"
 
 #include <DRAMPower/standards/ddr5/DDR5.h>
