+++ src/DRAMPower/DRAMPower/standards/lpddr4/core_calculation_LPDDR4.cpp
@@ -1,4 +1,4 @@
-#include "calculation_LPDDR4.h"
+#include "core_calculation_LPDDR4.h"
 
 #include <DRAMPower/standards/lpddr4/LPDDR4.h>
